- series completed

+++ app/Http/Controllers/PostsController.php
@@ -4,9 +4,20 @@ namespace App\Http\Controllers;
 
 use Illuminate\Http\Request;
 use App\Post;
+use Illuminate\Support\Facades\Storage;
 
 class PostsController extends Controller
 {
+    /**
+     * Create a new controller instance.
+     *
+     * @return void
+     */
+    public function __construct()
+    {
+        $this->middleware('auth', ['except' => ['index', 'show']]);
+    }
+
     /**
      * Display a listing of the resource.
      *
@@ -39,14 +50,34 @@ class PostsController extends Controller
     {
         $this->validate($request, [
             'title' => 'required',
-            'body' => 'required'
+            'body' => 'required',
+            'cover_image' => 'image|nullable|max:1999'
         ]);
+
+        // handle file upload
+        if ($request->hasFile('cover_image')) {
+            // get just filename
+            $filename = pathinfo($request->file('cover_image')->getClientOriginalName(), PATHINFO_FILENAME);
+
+            // get just extension
+            $extension = $request->file('cover_image')->getClientOriginalExtension();
+
+            // filename to store
+            $fileNameToStore = $filename . '_' . time() . '.' . $extension;
+
+            // upload
+            $path = $request->file('cover_image')->storeAs('public/cover_images', $fileNameToStore);
+        }
+        else {
+            $fileNameToStore = 'noimage.jpg';
+        }
 
         // create post
         $post = new Post();
         $post->title = $request->input('title');
         $post->body = $request->input('body');
         $post->user_id = auth()->user()->id;
+        $post->cover_image = $fileNameToStore;
 
         $post->save();
 
@@ -74,6 +105,12 @@ class PostsController extends Controller
     public function edit($id)
     {
         $post = Post::find($id);
+
+        // check for corect user
+        if (auth()->user()->id != $post->user_id) {
+            return redirect('/posts')->with('error', 'Unauthorized page');
+        }
+
         return view('posts.edit')->with('post', $post);
     }
 
@@ -88,13 +125,34 @@ class PostsController extends Controller
     {
         $this->validate($request, [
             'title' => 'required',
-            'body' => 'required'
+            'body' => 'required',
+            'cover_image' => 'image|nullable|max:1999'
         ]);
+
+        // handle file upload
+        if ($request->hasFile('cover_image')) {
+            // get just filename
+            $filename = pathinfo($request->file('cover_image')->getClientOriginalName(), PATHINFO_FILENAME);
+
+            // get just extension
+            $extension = $request->file('cover_image')->getClientOriginalExtension();
+
+            // filename to store
+            $fileNameToStore = $filename . '_' . time() . '.' . $extension;
+
+            // upload
+            $path = $request->file('cover_image')->storeAs('public/cover_images', $fileNameToStore);
+        }
 
         // create post
         $post = Post::find($id);
         $post->title = $request->input('title');
         $post->body = $request->input('body');
+
+        if ($request->hasFile('cover_image')) {
+            $post->cover_image = $fileNameToStore;
+        }
+
         $post->save();
 
         return redirect('/posts')->with('success', 'Post updated');
@@ -109,6 +167,17 @@ class PostsController extends Controller
     public function destroy($id)
     {
         $post = Post::find($id);
+
+        // check for corect user
+        if (auth()->user()->id != $post->user_id) {
+            return redirect('/posts')->with('error', 'Unauthorized page');
+        }
+
+        if ($post->cover_image != 'noimage.jpg') {
+            // delete image
+            Storage::delete('public/cover_images/' . $post->cover_image);
+        }
+
         $post->delete();
         return redirect('/posts')->with('success', 'Post deleted');
     }
